flagships/condo_board_decision_pack/pyproject.toml
[project]
name = "ks-cookbook-condo-board"
version = "0.1.0"
description = "CondoStack flagship — adjudicate a unit-owner architectural request with a board-ready decision memo citing CC&Rs, bylaws, ACC guidelines, reserve-fund impact, and historical minutes."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[mcp]>=0.0.14",
    "pydantic>=2.6",
    "python-docx>=1.1",
]

[project.scripts]
ks-cookbook-condo-board = "condo_board_decision.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/condo_board_decision"]
